refactor(type): Extract the compile-time type metadata into its own header
Reaching TypeKind, TypeTraits or SimpleTypeTrait currently means parsing all
~2400 lines of Type.h, which brings folly/Random.h,
folly/container/F14Set.h, common/serialization/Serializable.h, glog, gflags,
<iomanip> and the virtual Type hierarchy with it. 232 files include Type.h
directly, and many want nothing more than the kind enum or a trait off it.

The block is cleanly separable. Two groups move to a new velox/type/TypeKind.h:

1. Lines 73-380: the TypeKind enum, the 18 TypeTraits specializations,
   UnknownValue, is_string_kind / is_nested_kind, and forward declarations of
   the Type impl classes. These are pure constexpr/POD and depend on none of
   Type.h's 27 includes. The duplicate `using int128_t = __int128_t;` at
   Type.h:53 moves along with them; HugeInt.h declares the same alias
   independently.

2. The SimpleTypeTrait primary template and its ten primitive specializations.
   SimpleTypeTrait maps a C++ type to the TypeTraits of the logical type it
   denotes, which is how simple-function registration derives a signature from
   its template arguments -- TypeAnalysis reads SimpleTypeTrait<T>::name for
   each argument and return type, and FunctionSignatureBuilder assembles them.
   Every specialization is `: public TypeTraits<TypeKind::X> {}` with an empty
   body, so once group 1 has moved they depend on nothing else.

The trait stays split across two headers, as it already is: the primitive half
moves here, and the specializations for signature tag types (Varchar, Map,
Array, Row, ...) remain in SimpleFunctionApi.h beside the tags they describe,
extending the same primary template.

TypeKind.h's only dependencies are <compare>, <cstdint>, <limits>, <memory>,
<string> and common/EnumDeclare.h, plus declarations of StringView and
Timestamp -- TypeTraits names those two solely to form type aliases, so
incomplete types suffice.

Type.h includes the new header, so all existing includers are unaffected and no
call site changes. Consumers that want the metadata without the hierarchy
behind it -- DecimalUtil.h, VectorTypeUtils.h, SimpleAggregateAdapter.h,
VectorReaders.h among them -- can now reach it directly.

VELOX_FLUENT_CAST stays in Type.h: it is defined, used, and #undef'd within the
Type class it generates members for, and expands to TypeTraits<KIND>::ImplType,
which Type.h still sees through this header.

The overview comment above the extracted block moves with it. It introduced the
TypeKind enum that directly followed, so after the extraction it described
nothing, and its claim that "this file has multiple C++ type definitions ... -
TypeKind" was no longer true. The TypeKind entry now points at its new home,
and TypeKind.h carries a header comment naming the same type set, so the two
files reference each other rather than each describing the split.

Pure extraction. No behavior change, no API change.

Verified: full velox build clean; TypeKind.h and Type.h each compile
standalone; velox_type_test 288, velox_functions_test 1926,
velox_functions_spark_test 707, velox_expression_test 1636 -- 4557 tests, all
passing.
